+++ zetaclient/chains/base/observer.go
@@ -13,9 +13,8 @@ import (
 	"github.com/rs/zerolog"
 
 	"github.com/zeta-chain/node/pkg/chains"
-	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 	observertypes "github.com/zeta-chain/node/x/observer/types"
-	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
+	"github.com/zeta-chain/node/zetaclient/chains/tssrepo"
 	"github.com/zeta-chain/node/zetaclient/chains/zrepo"
 	"github.com/zeta-chain/node/zetaclient/db"
 	"github.com/zeta-chain/node/zetaclient/logs"
@@ -51,7 +50,7 @@ type Observer struct {
 	zetaRepo *zrepo.ZetaRepo
 
 	// tssSigner is the TSS signer
-	tssSigner interfaces.TSSSigner
+	tssSigner tssrepo.TSSClient
 
 	// lastBlock is the last block height of the observed chain
 	lastBlock uint64
@@ -66,7 +65,7 @@ type Observer struct {
 
 	// internalInboundTrackers stores trackers for inbounds that failed to vote on due to broadcasting error (e.g. tx dropped)
 	// the contents of the map may vary from observer to observer, depending on individual situation
-	internalInboundTrackers map[string]crosschaintypes.InboundTracker
+	internalInboundTrackers map[string]*InternalInboundTracker
 
 	// db is the database to persist data
 	db *db.DB
@@ -91,7 +90,7 @@ func NewObserver(
 	chain chains.Chain,
 	chainParams observertypes.ChainParams,
 	zetaRepo *zrepo.ZetaRepo,
-	tssSigner interfaces.TSSSigner,
+	tssSigner tssrepo.TSSClient,
 	blockCacheSize int,
 	ts *metrics.TelemetryServer,
 	database *db.DB,
@@ -113,7 +112,7 @@ func NewObserver(
 		ts:                      ts,
 		db:                      database,
 		blockCache:              blockCache,
-		internalInboundTrackers: make(map[string]crosschaintypes.InboundTracker),
+		internalInboundTrackers: make(map[string]*InternalInboundTracker),
 		mu:                      &sync.Mutex{},
 		logger:                  newObserverLogger(chain, logger),
 		stop:                    make(chan struct{}),
@@ -191,7 +190,7 @@ func (ob *Observer) ZetaRepo() *zrepo.ZetaRepo {
 }
 
 // TSS returns the tss signer for the observer.
-func (ob *Observer) TSS() interfaces.TSSSigner {
+func (ob *Observer) TSS() tssrepo.TSSClient {
 	return ob.tssSigner
 }