+++ zetaclient/chains/zrepo/zrepo.go
@@ -26,16 +26,23 @@ import (
 	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
 
-// MonitoringErrHandlerRoutineTimeout is the timeout for the handleMonitoring routine that waits for an error from the monitorVote channel
-const monitoringErrHandlerRoutineTimeout = 5 * time.Minute
+const (
+	// inboundVoteTimeout is the timeout for a single inbound vote.
+	inboundVoteTimeout = 10 * time.Minute
+
+	// monitoringErrWatcherTimeout is the timeout for the monitoring error watcher routine that waits
+	// for an error from the monitorVote channel. It is slightly longer than the inbound vote timeout
+	// to be able to catch the vote tx errors, including the timeout error.
+	monitoringErrWatcherTimeout = inboundVoteTimeout + 30*time.Second
+)
 
 // MonitoringErrorWatcher is the function type for watching inbound vote monitoring errors.
-type MonitoringErrorWatcher func(ctx context.Context, monitorErrCh <-chan zetaerrors.ErrTxMonitor, zetaTxHash string)
+type MonitoringErrorWatcher func(_ context.Context, _ <-chan zetaerrors.ErrTxMonitor, zhash string)
 
 // ZetaRepo implements the Repository pattern by wrapping a zetacore client.
 // Each chain module must instantiate its own ZetaRepo.
 type ZetaRepo struct {
-	client ZetacoreClientRepo
+	client ZetacoreClient
 
 	connectedChain chains.Chain
 
@@ -44,6 +51,9 @@ type ZetaRepo struct {
 
 // New constructs a new ZetaRepo object.
 func New(client ZetacoreClient, connectedChain chains.Chain, clientMode mode.ClientMode) *ZetaRepo {
+	if client == nil {
+		return nil
+	}
 	return &ZetaRepo{client, connectedChain, clientMode}
 }
 
@@ -108,7 +118,10 @@ func (repo *ZetaRepo) GetBTCTSSAddress(ctx context.Context) (string, error) {
 	return address, nil
 }
 
-func (repo *ZetaRepo) HasVoted(ctx context.Context, ballotIndex string, voterAddress string) (bool, error) {
+func (repo *ZetaRepo) HasVoted(ctx context.Context,
+	ballotIndex string,
+	voterAddress string,
+) (bool, error) {
 	return repo.client.HasVoted(ctx, ballotIndex, voterAddress)
 }
 
@@ -217,14 +230,23 @@ func (repo *ZetaRepo) VoteInbound(ctx context.Context, logger zerolog.Logger,
 		return "", nil
 	}
 
-	// ctxWithTimeout is a context with timeout used for monitoring the vote transaction
-	// Note: the canceller is not used because we want to allow the goroutines to run until they time out
-	ctxWithTimeout, _ := zctx.CopyWithTimeout(ctx, context.Background(), monitoringErrHandlerRoutineTimeout)
+	// ctxInboundVote is a context with timeout used for the inbound vote.
+	// If this is a finalizing vote, the timeout is total time allowed for 500K vote and 7M vote.
+	// The canceller is ignored as we want to allow the goroutine to run until it times out.
+	ctxInboundVote, _ := zctx.CopyWithTimeout(ctx, context.Background(), inboundVoteTimeout)
+
+	// ctxMonitorErrWatcher is a context with timeout used for monitor error watcher routine.
+	// The canceller is ignored as we want to allow the goroutine to run until it times out.
+	ctxMonitorErrWatcher, _ := zctx.CopyWithTimeout(ctx, context.Background(), monitoringErrWatcherTimeout)
 
 	// Post vote to zetacore.
-	const gasLimit = zetacore.PostVoteInboundGasLimit
 	monitorErrCh := make(chan zetaerrors.ErrTxMonitor, 1)
-	zhash, ballot, err := repo.client.PostVoteInbound(ctxWithTimeout, gasLimit, retryGasLimit, msg, monitorErrCh)
+	zhash, ballot, err := repo.client.PostVoteInbound(ctxInboundVote,
+		zetacore.PostVoteInboundGasLimit,
+		retryGasLimit,
+		msg,
+		monitorErrCh,
+	)
 	if err != nil {
 		err = newClientError(ErrClientVoteInbound, err)
 		logger.Error().Err(err).Send()
@@ -237,10 +259,10 @@ func (repo *ZetaRepo) VoteInbound(ctx context.Context, logger zerolog.Logger,
 	} else {
 		logger.Info().Str(logs.FieldZetaTx, zhash).Msg("posted inbound vote")
 
-		// watch for monitoring error for this vote
+		// Watch for monitoring errors for this vote.
 		if monitorErrWatcher != nil {
 			go func() {
-				monitorErrWatcher(ctxWithTimeout, monitorErrCh, zhash)
+				monitorErrWatcher(ctxMonitorErrWatcher, monitorErrCh, zhash)
 			}()
 		}
 	}
@@ -320,6 +342,11 @@ func (repo *ZetaRepo) GetForeignCoinsFromAsset(ctx context.Context,
 	return &coins, nil
 }
 
+func (repo *ZetaRepo) CCTXExists(ctx context.Context, hash string) (bool, error) {
+	f := repo.client.GetCctxByHash
+	return exists(ctx, hash, f, grpccodes.InvalidArgument, ErrClientGetCCTXByHash)
+}
+
 // ------------------------------------------------------------------------------------------------
 // Auxiliary functions
 // ------------------------------------------------------------------------------------------------
@@ -342,7 +369,7 @@ func checkCode(err error, code grpccodes.Code) error {
 	return nil
 }
 
-// exists is the generic function used by cctxExists and ballotExists to check for CCTXs and
+// exists is the generic function used by CCTXExists and ballotExists to check for CCTXs and
 // ballots.
 func exists[T any](ctx context.Context,
 	hashOrID string, // hash of a CCTX or the ID of a ballot
@@ -359,11 +386,6 @@ func exists[T any](ctx context.Context,
 		return false, nil
 	}
 	return res != nil, nil
-}
-
-func (repo *ZetaRepo) CCTXExists(ctx context.Context, hash string) (bool, error) {
-	f := repo.client.GetCctxByHash
-	return exists(ctx, hash, f, grpccodes.InvalidArgument, ErrClientGetCCTXByHash)
 }
 
 func (repo *ZetaRepo) ballotExists(ctx context.Context, id string) (bool, error) {